+++ doc/whatsnew.rst
@@ -54,6 +54,7 @@ Internal changes
 
 - Move dataset accessor to :mod:`hyoga.core.accessor`.
 - Move hyoga configuration to :mod:`hyoga.core.config`.
+- Move local input methods to :mod: `hyoga.open.local`.
 - Move dataset plot methods to :mod:`hyoga.plot.datasets`.